Re-Introduce Code Cleanup for New Code
