+++ ClassicMode.xaml
@@ -2,7 +2,7 @@
     x:Class="Files.ClassicMode"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local1="using:ItemListPresenter"
+    xmlns:local1="using:Files.Filesystem"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:UILib="using:Microsoft.UI.Xaml.Controls"
