+++ ClassicMode.xaml.cs
@@ -1,10 +1,9 @@
-﻿using ItemListPresenter;
-using Windows.ApplicationModel.Core;
+﻿using Windows.ApplicationModel.Core;
 using Windows.UI;
 using Windows.UI.ViewManagement;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
-
+using Files.Filesystem;
 
 namespace Files
 {
