+++ GenericFileBrowser.xaml.cs
@@ -1,7 +1,4 @@
-﻿using Interacts;
-using ItemListPresenter;
-using Microsoft.Toolkit.Uwp.UI.Controls;
-using Navigation;
+﻿using Microsoft.Toolkit.Uwp.UI.Controls;
 using System;
 using System.ComponentModel;
 using System.Diagnostics;
@@ -16,6 +13,8 @@ using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
 using Windows.UI.Xaml.Media.Animation;
 using Windows.UI.Xaml.Navigation;
+using Files.Filesystem;
+using Files.Navigation;
 
 namespace Files
 {
@@ -116,7 +115,7 @@ namespace Files
         }
 
         public static UniversalPath p = new UniversalPath();
-        public static UniversalPath P { get { return GenericFileBrowser.p; } }
+        public static UniversalPath P { get { return p; } }
         
         protected override void OnNavigatedTo(NavigationEventArgs eventArgs)
         {
@@ -192,22 +191,22 @@ namespace Files
 
         private async void AllView_CellEditEnded(object sender, DataGridCellEditEndedEventArgs e)
         {
-            var NewCellText = (GenericFileBrowser.data.SelectedItem as ListedItem).FileName;
-            var SelectedItem = ItemViewModel.FilesAndFolders[e.Row.GetIndex()];
-            if(SelectedItem.FileExtension == "Folder")
+            var newCellText = (GenericFileBrowser.data.SelectedItem as ListedItem)?.FileName;
+            var selectedItem = ItemViewModel.FilesAndFolders[e.Row.GetIndex()];
+            if(selectedItem.FileExtension == "Folder")
             {
-                StorageFolder FolderToRename = await StorageFolder.GetFolderFromPathAsync(SelectedItem.FilePath);
-                if(FolderToRename.Name != NewCellText)
+                StorageFolder FolderToRename = await StorageFolder.GetFolderFromPathAsync(selectedItem.FilePath);
+                if(FolderToRename.Name != newCellText)
                 {
-                    await FolderToRename.RenameAsync(NewCellText);
+                    await FolderToRename.RenameAsync(newCellText);
                 }
             }
             else
             {
-                StorageFile FileToRename = await StorageFile.GetFileFromPathAsync(SelectedItem.FilePath);
-                if (FileToRename.Name != NewCellText)
+                StorageFile fileToRename = await StorageFile.GetFileFromPathAsync(selectedItem.FilePath);
+                if (fileToRename.Name != newCellText)
                 {
-                    await FileToRename.RenameAsync(NewCellText);
+                    await fileToRename.RenameAsync(newCellText);
                 }
             }
             Navigation.NavigationActions.Refresh_Click(null, null);
