+++ Interacts/ItemInteractions.cs
@@ -1,0 +1,33 @@
+﻿using System.ComponentModel;
+
+namespace Files.Interacts
+{
+    public class PasteState : INotifyPropertyChanged
+    {
+        public bool _isEnabled;
+        public bool isEnabled
+        {
+            get
+            {
+                return _isEnabled;
+            }
+
+            set
+            {
+                if (value != _isEnabled)
+                {
+                    _isEnabled = value;
+                    NotifyPropertyChanged("isEnabled");
+                }
+            }
+        }
+        public event PropertyChangedEventHandler PropertyChanged;
+
+        private void NotifyPropertyChanged(string info)
+        {
+            PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(info));
+        }
+    }
+}
+
+
