+++ LocationsList.cs
@@ -12,12 +12,9 @@ namespace Locations
 
     public class ItemLoader
     {
-        
         public ItemLoader()
         {
-
         }
-
         public static List<LocationItem> itemsAdded = new List<LocationItem>();
         public static void DisplayItems()
         {
@@ -26,9 +23,6 @@ namespace Locations
             itemsAdded.Add(new LocationItem() { ImageSource = "Assets/Cards/Gradients/Orange.png", Icon = "\xEB9F", Text = "Pictures" });
             itemsAdded.Add(new LocationItem() { ImageSource = "Assets/Cards/Gradients/Pink.png", Icon = "\xEC4F", Text = "Music" });
             itemsAdded.Add(new LocationItem() { ImageSource = "Assets/Cards/Gradients/Red.png", Icon = "\xE8B2", Text = "Videos" });
-
-
-
         }
     }
 }
