+++ YourHome.xaml.cs
@@ -1,5 +1,4 @@
-﻿using ItemListPresenter;
-using Microsoft.Toolkit.Uwp.UI.Controls;
+﻿using Microsoft.Toolkit.Uwp.UI.Controls;
 using System;
 using System.Diagnostics;
 using System.Drawing;
@@ -7,7 +6,7 @@ using Windows.Graphics.Display;
 using Windows.UI.ViewManagement;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
-
+using Files.Filesystem;
 
 namespace Files
 {
